% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andarl.R
\name{ad.test}
\alias{ad.test}
\title{Anderson-Darling Test of Goodness-of-Fit}
\usage{
ad.test(x, null = "punif", ..., nullname)
}
\arguments{
\item{x}{Numeric vector of data values.}

\item{null}{A function, or a character string giving the name of a function, to compute the cumulative
distribution function for the null distribution}

\item{...}{Additional arguments for the cumulative distribution function.}

\item{nullname}{Optional character string describing the null distribution. The default is "uniform distribution".}
}
\value{
An object of class "htest" representing the result of the hypothesis test
}
\description{
Performs the Anderson-Darling test of goodness-of-fit to a specified continuous univariate proba
bility distribution
}
\details{
This command performs the Anderson-Darling test of goodness-of-fit to the distribution specified
by the argument null. It is assumed that the values in x are independent and identically distributed
random values, with some cumulative distribution function F. The null hypothesis is that F is the
function specified by the argument null, while the alternative hypothesis is that F is some other
function.
}
\examples{

x <- rnorm(10, mean=2, sd=1)
ad.test(x, "pnorm", mean=2, sd=1)

}
\references{
Anderson, T.W. and Darling, D.A. (1952) Asymptotic theory of certain ’goodness-of-fit’ criteria
based on stochastic processes. Annals of Mathematical Statistics 23, 193–212.

Anderson, T.W. and Darling, D.A. (1954) A test of goodness of fit. Journal of the American Statistical Association 49, 765–769.

Marsaglia, G. and Marsaglia, J. (2004) Evaluating the Anderson-Darling Distribution. Journal of Statistical Software 9 (2), 1–5.
February 2004. http://www.jstatsoft.org/v09/i02
}
\author{
Original C code by George Marsaglia and John Marsaglia. R interface by Adrian Baddeley
}
