% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transformation.R
\name{box.cox}
\alias{box.cox}
\title{Box cox transformation}
\usage{
box.cox(s, lambda = 0)
}
\arguments{
\item{s}{vector with data}

\item{lambda}{lambda parameter of box-cox tranformation}
}
\value{
Box-Cox transformation of input vector with lambda parameter
}
\description{
Box cox transformation
}
