% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transformation.R
\name{boxcox}
\alias{boxcox}
\title{boxcox transformation and evaluation}
\usage{
boxcox(x = x, lambda = NULL, plot = F, ks.test = F, ...)
}
\arguments{
\item{x}{vector}

\item{lambda}{(Optional) lambda, if NULL, lambda is optimized}

\item{plot}{(Optional) Logical, plot qqplots}

\item{ks.test}{(Optional) Logical, kolmogorov-smirnov test}

\item{...}{additional arguments fo lambda.box}
}
\value{
Compute the box-cox transformation based of a vector. Plot and ks.test argument can be set to TRUE to asses the normality of the new vector
}
\description{
boxcox transformation and evaluation
}
\seealso{
lambda.box
}
