% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_opt.R
\name{cer_optim}
\alias{cer_optim}
\title{CeR optimization}
\usage{
cer_optim(A = A, B = B, C = C, D = D, PD = PD, LGD = LGD,
  weight = weight, EAD = EAD, CORR = CORR, UL = UL, CONTRI = CONTRI,
  variation = 0.05, maxiter = 10000, parallel = F, tol = 1e-09, ...)
}
\arguments{
\item{A}{A}

\item{B}{B}

\item{C}{C}

\item{D}{D}

\item{PD}{PD}

\item{LGD}{LGD}

\item{weight}{N_client}

\item{EAD}{EAD}

\item{CORR}{CORR}

\item{UL}{UL}

\item{CONTRI}{Conttribution}

\item{variation}{range of parameter variation}

\item{maxiter}{Numeric, the maximum number of iterations to run}

\item{parallel}{a logical argument specifying if parallel computing
should be used (TRUE) or not (FALSE, default) for evaluating the
fitness function. This argument could also be used to specify the
number of cores to employ; by default, this is taken from detectCores.
Finally, the functionality of parallelization depends on system OS: on
Windows only 'snow' type functionality is available, while on
Unix/Linux/Mac OSX both 'snow' and 'multicore' (default) functionalities
are available}

\item{tol}{the desired accuracy.}

\item{...}{additional named or unnamed arguments to be passed to ga}
}
\description{
CeR optimization
}
