% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_opt.R
\name{cor_optim}
\alias{cor_optim}
\title{Global & Local factor optimization}
\usage{
cor_optim(map = map, hist = hist, CD = CD, lim = 1, maxiter = 10000,
  parallel = F, ...)
}
\arguments{
\item{map}{Dataframe with the correspondence}

\item{hist}{Matrix with historical correlation}

\item{CD}{Matrix with credit drivers correlation}

\item{lim}{Numeric it acts as a limit of FG ^ 2 + FL ^ 2, by default is 0.6}

\item{maxiter}{Numeric the maximum number of iterations to run}

\item{parallel}{Logical, argument specifying if parallel computing should be used
(TRUE) or not (FALSE, default) for evaluating the fitness function. This argument
could also be used to specify the number of cores to employ; by default, this is
taken from detectCores. Finally, the functionality of parallelization depends on
system OS: on Windows only 'snow' type functionality is available, while on Unix/Linux/Mac
OSX both 'snow' and 'multicore' (default) functionalities are available.}

\item{...}{Arguments to be passed to ga function}
}
\description{
Global & Local factor optimization
}
\examples{
x = cor_optim(map = map,
              hist = hist,
              CD = CD,
              lim = 1,
              maxiter = 500,
              run = 10)

x = cor_optim(map = map,
              hist = hist,
              CD = CD,
              lim = 1,
              maxiter = 500,
              run = 10,
              suggestions = get_suggestions(x))

}
