% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_stress.R
\name{cor_stress}
\alias{cor_stress}
\title{Stressed correlation}
\usage{
cor_stress(x, y = NULL, period = period, abs = F, use = "everything",
  method = c("pearson", "kendall", "spearman"), ...)
}
\arguments{
\item{x}{a numeric vector, matrix or data frame.}

\item{y}{NULL (default) or a vector, matrix or data frame with compatible
dimensions to x. The default is equivalent to y = x (but more efficient).}

\item{period}{numeric number of periods to use}

\item{abs}{logical. Should absolute values be used}

\item{use}{an optional character string giving a method for computing
covariances in the presence of missing values. This must be (an
abbreviation of) one of the strings "everything", "all.obs",
"complete.obs", "na.or.complete", or "pairwise.complete.obs"}

\item{method}{a character string indicating which correlation coefficient
(or covariance) is to be computed. One of "pearson" (default), "kendall",
or "spearman": can be abbreviated.}

\item{...}{}
}
\value{
Return the stressed correlation of x and y in the selected period
if these are vectors. If x and y are matrices then the correlations
between the columns of x and the columns of y are returned
}
\description{
Stressed correlation
}
