% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_punctuation_spacing.R
\name{correct_punctuation_spacing}
\alias{correct_punctuation_spacing}
\title{\code{correct_punctuation_spacing()}}
\usage{
correct_punctuation_spacing(name, punctuation = c(",", ";", ":"))
}
\arguments{
\item{name}{A character vector containing one or more strings in which to correct spacing around punctuation marks.}

\item{punctuation}{A character vector containing punctuation marks to correct spacing around; \code{c(',', ';', ':')} is default.}
}
\value{
A character vector.
}
\description{
Eliminates spaces before and ensures spaces after designated punctuation marks.
}
\details{
\code{correct_punctuation_spacing()} eliminates spaces before and ensures spaces after punctuation marks declared in
the \code{punctuation} argument. For example, "SMITH ,JANE" would be corrected to "SMITH, JANE" (assuming that a comma
is included among the designated punctuation marks via the \code{punctuation} argument. The \code{punctuation} argument has
a default value of \code{c(',', ';', ':')}. Before correcting spacing around the designated marks, the function checks whether
each entry in the \code{punctuation} argument is actually a punctuation mark; it will remove any that are not punctuation marks
and will alert the user via a warning.
}
\examples{
correct_punctuation_spacing(c('SMITH ,JANE', 'SMITH,JANE', 'SMITH , JANE'),
                            punctuation = c(',', ';', ':', 'ABC'))

}
