% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcorlag.R
\name{corscale}
\alias{corscale}
\title{corscale}
\usage{
corscale(DATA = DATA, ref = ref, variables = variables, lag.max = 6,
  allow.negative = T, max = T)
}
\arguments{
\item{DATA}{data.frame with the data}

\item{ref}{a string with the reference variable}

\item{variables}{variables to change}

\item{lag.max}{maximum number of lags to asses}

\item{allow.negative}{(Optional) logical, if negative lags are allowed, by default TRUE}

\item{max}{(Optional) logical, Should maximum be returned, in other case minimum is returned}
}
\value{
Return a dataframe with the variables lagged or delayed with the max correlation to reference variable
}
\description{
corscale
}
