% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcorlag.R
\name{cutcor}
\alias{cutcor}
\title{cutcor}
\usage{
cutcor(data = data, FG = FG)
}
\arguments{
\item{data}{dataframe}

\item{FG}{character}
}
\value{
return a dataframe with more correlation between the global factor and the distance to default vector
}
\description{
cutcor
}
