% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_cor.R
\name{diss_cor}
\alias{diss_cor}
\title{Dissimilarity Index Combining Temporal Correlation and Raw Values}
\usage{
diss_cor(SERIES, ...)
}
\arguments{
\item{SERIES}{data}

\item{...}{beta = NULL}
}
\value{
The computed distance.
}
\description{
Computes an adaptive dissimilarity index between two time series that covers both dissimilarity on
raw values and dissimilarity on temporal correlation behaviors.
}
\details{
¡¡FILL!!
}
\references{
Chouakria-Douzal, A. and Nagabhushan P. N. (2007) Adaptive dissimilarity index for measuring
time series proximity. Adv. Data Anal. Classif., 1(1), 5–21.

Montero, P and Vilar, J.A. (2014) TSclust: An R Package for Time Series Clustering. Journal of
Statistical Software, 62(1), 1-43. http://www.jstatsoft.org/v62/i01/.
}
