% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ec_norm_plot.R
\name{ec_norm_plot}
\alias{ec_norm_plot}
\title{ec_norm_plot}
\usage{
ec_norm_plot(variable, indicator, lan = "english")
}
\arguments{
\item{variable}{Numeric vector of raw variable values.}

\item{indicator}{Numeric vector of normalised indicator values, typically
produced by \code{ec_normalise_dev()}.}

\item{lan}{Character. Plot language. One of \code{"english"} or \code{"norsk"}.
Defaults to \code{"english"}.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot normalised ecological indicator values against the raw variable values
}
\details{
Creates a scatter plot comparing raw variable values and corresponding
normalised indicator values produced by \code{ec_normalise_dev()}.

The plot is intended as a simple diagnostic and visualisation tool for
evaluating ecological normalisation functions.

Points are coloured according to the indicator value using a red-to-green
colour gradient.

The function returns a \code{ggplot2} object and can therefore be further modified
using standard \code{ggplot2} syntax.
}
\examples{
x <- seq(0, 1000, by = 2)

ind <- ec_normalise(
  variable = x,
  x0 = 100,
  x100 = 900,
  x60 = 200
)

ec_norm_plot(
  variable = x,
  indicator = ind,
  lan = "norsk"
)

ind <- ec_normalise(
  variable = x,
  x0 = 50,
  x100 = rnorm(n = length(x), 400, 50),
  x0h = 900,
  fun = "sigmoid"
)

ec_norm_plot(
  variable = x,
  indicator = ind
)
}
