% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_shorthand.R
\name{expand_shorthand}
\alias{expand_shorthand}
\title{\code{expand_shorthand()}}
\usage{
expand_shorthand(to_expand, type = "housing", case_out = "upper")
}
\arguments{
\item{to_expand}{A vector containing one or more strings in which to expand shorthands/acronyms. Character or character-coercible.}

\item{type}{The "genre" of expansions to perform; currently, only housing-related expansions are provided.}

\item{case_out}{Should the \code{to_expand} vector, with expansions, be returned in uppercase or lowercase? Uppercase is default.}
}
\value{
A character vector.
}
\description{
Expands shorthands and acronyms in a vector of strings based on a table(s) included in the package.
}
\details{
\code{expand_shorthand()} expands shorthands and acronyms in a vector of strings according to table(s) included in the package. The function currently
offers expansions for housing-related terms, with more to come.
}
\examples{
example_names <- c('GREENE APTS', 'BLU MGMT', 'REDD PROPERTY MGT')
expand_shorthand(to_expand = example_names, type = 'housing', case_out = 'upper')

}
