% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saddlepoint.R
\name{fit_saddlepoint}
\alias{fit_saddlepoint}
\title{Saddle-point aproximation}
\usage{
fit_saddlepoint(x = x, N_SIMUL = 10000, n_days = 1, nm = 6)
}
\arguments{
\item{x}{vector}

\item{N_SIMUL}{(Optional) Number of simulations. Default 10000}

\item{n_days}{(Optional) Number of periods. Default 1}

\item{nm}{number of momements to use. Must be 2, 4 or 6}
}
\description{
fit_saddlepoint
}
