% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_skew_t.R
\name{fit_skewt}
\alias{fit_skewt}
\title{Moment estimation of the skew-t distribution}
\usage{
st = fit_skewt(x)
st
plot(st)
summary(st)
qst(0.99, df = get_param(st))
ks_test(st)
qqPlot(st)
}
\arguments{
\item{x}{Numeric vector to fit}

\item{n_days}{Number of days, to convolve the distribution}
}
\value{
This functions returns the estimated parameters of skew-t using the moment method and plot the observed and the fitted distribution

* skewtpar functions return a list with the following values:

* value, value of the function to minimize

* shape shape parameter to skew-t distribution

* df, nu parameter to skew-t distribution

* escala, omega parameter to skew-t distribution

* location, xi parameter to skew-t distribution

* convergence, logical value indicating whether the optimization has converged
or not. 1 indicates successful convergence and 0 no convergence. A successful
convergence is considered when the value of objetive the funcion is less
than 0.0001
}
\description{
Moment estimation of the skew-t distribution
}
\examples{

st = fit_skewt(x)
st
plot(st)
summary(st)
qst(0.99, df = get_param(st))
ks_test(st)
qqPlot(st)

}
\references{
Technical Note No. 2. Skew t Distribution of Economic Capital Models Department Methodology & Model Development Area
}
