% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regress.R
\name{flresi}
\alias{flresi}
\title{Get the residuals from a regression}
\usage{
flresi(y = y, x.var = x.var, scale = T)
}
\arguments{
\item{y}{dataframe or vector}

\item{x.var}{character variable, representing the dependent variable}

\item{scale}{logical, it the result should be standardie}
}
\value{
This function perform regressions for every variable inside 'y',
'y[i] ~ x.var' and get the residuals 'y[i] - b0 * b1 x.var'

In this way we get orthogonal vectors to 'x.var'
}
\description{
Get the residuals from a regression
}
