% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_opt.R
\name{Fx_class}
\alias{Fx_class}
\title{Title}
\usage{
Fx_class(hist = hist, CD = CD, map = map, sensitivities = sensitivities,
  lim = 1, FX = "FX", maxiter = 10000, parallel = F, minFG = 0,
  seed = 1, ...)
}
\arguments{
\item{hist}{Matrix with historical correlation}

\item{CD}{Matrix with credit drivers correlation}

\item{map}{Dataframe with the correspondence}

\item{sensitivities}{DataFrame. with sensitivities}

\item{lim}{Numeric, limit of the sum squared of the factors, by default 1}

\item{FX}{Character, name of the label of the FX}

\item{parallel}{Logical, argument specifying if parallel computing should be used
(TRUE) or not (FALSE, default) for evaluating the fitness function. This argument
could also be used to specify the number of cores to employ; by default, this is
taken from detectCores. Finally, the functionality of parallelization depends on
system OS: on Windows only 'snow' type functionality is available, while on Unix/Linux/Mac
OSX both 'snow' and 'multicore' (default) functionalities are available.}

\item{...}{Arguments to be passed to the optimization function}

\item{lim}{Numeric, limit of the sum squared of the factors, by default 1}
}
\description{
Title
}
