% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_non_residential_names.R
\name{identify_non_residential_names}
\alias{identify_non_residential_names}
\title{\code{identify_non_residential_names()}}
\usage{
identify_non_residential_names(name, use_alt_regex = FALSE)
}
\arguments{
\item{name}{A vector containing one or more strings to identify as non-residential or residential. Character or character-coercible.}

\item{use_alt_regex}{Default (safe) is \code{FALSE}; if \code{TRUE}, the function will request a \code{source()}-able connection
to a file that, when executed, will return the regex to use to determine whether each element of \code{name} is non-residential or residential.}
}
\value{
A logical vector.
}
\description{
Identifies non-residential names ("La Sucre Bakery" vs. "Jan La Sucre") based on an extensive regex pattern.
}
\details{
\code{identify_non_residential_names()} applies a regex pattern to a character vector containing one or more names, returning
\code{TRUE} and \code{FALSE} for each element. The regex pattern used by default is taken from the GitHub repo jacob-gg/non-person-regex.
A user can, optionally, use a different regex by setting the \code{use_alt_regex} argument to \code{TRUE} and declaring
a connection; this, however, is usually not desired.
}
\examples{
example_names <- c('JANE SMITH', 'JANE SMITH APARTMENTS', 'AMERICANO CAFE')
identify_non_residential_names(example_names)

}
