% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k1}
\alias{k1}
\title{The first raw moment}
\usage{
k1(x)
}
\arguments{
\item{loss}{vector with input}

\item{scale}{scale parameter}
}
\value{
The first raw moment is the mean
}
\description{
The first raw moment
}
\references{
https://en.wikipedia.org/wiki/Moment_(mathematics)
}
