% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k2}
\alias{k2}
\title{The second central moment}
\usage{
k2(x)
}
\arguments{
\item{loss}{vector with input}

\item{scale}{scale parameter}
}
\value{
The second central moment is the variance. Its positive square root is the standard deviation σ.
}
\description{
The second central moment
}
\references{
https://en.wikipedia.org/wiki/Moment_(mathematics)
}
