% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_retail.R
\name{k2_boost}
\alias{k2_boost}
\title{Boostraping of the second central moment}
\usage{
k2_boost(x = x, p = p, n_iter = 100, q = 0.9995, rnd = T)
}
\arguments{
\item{x}{a numeric vector}

\item{p}{fraction of numbers to select}

\item{n_iter}{number of iterations}

\item{q}{quantile to return from the distribution generated}

\item{rnd}{logical, if TRUE the number are randomly selected}
}
\value{
In each iteration a subset from the vector x is selected
for variance computation, in this way a distribution is obtained.
Then, the value returned will be the quantile selected from the variance
distribution
}
\description{
Boostraping of the second central moment
}
