% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k3}
\alias{k3}
\title{The third central moment}
\usage{
k3(x)
}
\arguments{
\item{loss}{vector with input}

\item{scale}{scale parameter}
}
\value{
The third central moment is the measure of the lopsidedness of the distribution; any symmetric distribution will have a third
central moment, if defined, of zero. The normalised third central moment is called the skewness, often Gamma. A distribution that is
skewed to the left (the tail of the distribution is longer on the left) will have a negative skewness. A distribution that is skewed
to the right (the tail of the distribution is longer on the right), will have a positive skewness.
For distributions that are not too different from the normal distribution, the median will be somewhere near MU − Gamma * Sigma / 6;
the mode about MU − Gamma * Sigma / 2.
}
\description{
The third central moment
}
\references{
https://en.wikipedia.org/wiki/Moment_(mathematics)
}
