% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{k4}
\alias{k4}
\title{The fourth central moment}
\usage{
k4(x)
}
\arguments{
\item{loss}{vector with input}

\item{scale}{scale parameter}
}
\value{
The fourth central moment is a measure of the heaviness of the tail of the distribution, compared to the normal distribution
of the same variance. Since it is the expectation of a fourth power, the fourth central moment, where defined, is always positive;
and except for a point distribution, it is always strictly positive. The fourth central moment of a normal distribution is 3 * Sigma 4.

The kurtosis K is defined to be the normalised fourth central moment minus 3 (Equivalently, as in the next section, it is the fourth
cumulant divided by the square of the variance). Some authorities do not subtract three, but it is usually more convenient to have the
normal distribution at the origin of coordinates.[4][5] If a distribution has heavy tails, the kurtosis will be high (sometimes called
leptokurtic); conversely, light-tailed distributions (for example, bounded distributions such as the uniform) have low kurtosis (sometimes
called platykurtic).

The kurtosis can be positive without limit, but K must be greater than or equal to Gamma * 2 − 2; equality only holds for binary distributions.
For unbounded skew distributions not too far from normal, K tends to be somewhere in the area of Gamma * 2 and 2 * Gamma * 2.

The inequality can be proven by considering

E[(T^2-aT-1)^2

where T = (X − μ)/σ. This is the expectation of a square, so it is non-negative for all a; however it is also a quadratic polynomial in a.
Its discriminant must be non-positive, which gives the required relationship.
}
\description{
The fourth central moment
}
\references{
https://en.wikipedia.org/wiki/Moment_(mathematics)
}
