% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transformation.R
\name{lambda.box}
\alias{lambda.box}
\title{Fit the lambda which optimize the loglikehood function}
\usage{
lambda.box(x = x, interval = c(-20, 20), tol = 1e-05)
}
\arguments{
\item{x}{vector or dataframe with data}

\item{interval}{a vector containing the end-points of the interval to be searched for the minimum., default c(-20,20)}

\item{tol}{the desired accuracy, default = 0.00001}
}
\value{
Return the lambda parameter which minimizes the loglikehood function
}
\description{
Fit the lambda which optimize the loglikehood function
}
