% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcorlag.R
\name{llcor}
\alias{llcor}
\title{Correlations bewteen lags}
\usage{
llcor(x = x, y = y, max.lag = 6, allow.negative = T)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{max.lag}{max.lag}

\item{allow.negative}{(Optional) logical, if negative lags are allowed, by default TRUE}
}
\value{
Return the correlation between x and y lags
}
\description{
Correlations bewteen lags
}
