% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transformation.R
\name{loglik.box}
\alias{loglik.box}
\title{loglik.box}
\usage{
loglik.box(x, lambda)
}
\arguments{
\item{x}{vector}

\item{lambda}{lambda}
}
\value{
Loglikehood function
}
\description{
loglik.box
}
