% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcorlag.R
\name{maxccf}
\alias{maxccf}
\title{maxccf}
\usage{
maxccf(x = x, y = y, lag.max = 6, allow.negative = T, max = T)
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{lag.max}{(Optional) maximum lag}

\item{allow.negative}{(Optional) logical, if negative lags are allowed, by default TRUE}

\item{max}{(Optional) logical, Should maximum be returned, in other case minimum is returned}
}
\value{
Return the lag with the maximum/minimum correlations bewteen two time series

if x is a data frame or matrix, u must use maxccfdf fucntion
}
\description{
maxccf
}
