% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeR.R
\name{RC_OUT}
\alias{RC_OUT}
\title{Contribucion de riesgo del CeR}
\usage{
RC_OUT(param = param, x = x)
}
\arguments{
\item{param}{vector with parameters c(A,B,C,D)}

\item{x}{dataframe con los datos}
}
\description{
Contribucion de riesgo del CeR
}
