% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates_df.R
\name{remove_duplicates_df}
\alias{remove_duplicates_df}
\title{\code{remove_duplicates_df()}}
\usage{
remove_duplicates_df(dat, column_names)
}
\arguments{
\item{dat}{A data frame.}

\item{column_names}{A character vector containing column names on which to base the assessment of row duplication.
If \code{NULL}, \code{column_names} is set to \code{colnames(dat)}.}
}
\value{
A data frame.
}
\description{
Removes duplicate rows from a data frame, determined based on a user-declared set of columns.
}
\details{
\code{remove_duplicates_df()} uses a user-declared set of variables to determine which rows in a data frame
are duplicates of others, and it removes all but the first instance of each duplicated row.
In effect, it's \code{dplyr::distinct(dat, var1, var2, ...)} with a bit more information in the output.
}
\examples{
d <- data.frame(year = c(1999, 1999, 1999, 1999),
                name = c('jane', 'jane', 'mark', 'erin'),
                misc = c(1, 2, 3, 4))
remove_duplicates_df(d, c('year', 'name'))

}
