% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox_transformation.R
\name{scaledf}
\alias{scaledf}
\title{scaledf}
\usage{
scaledf(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{dataframe}

\item{center}{either a logical value or a numeric vector of length equal to the number of columns of x.}

\item{scale}{either a logical value or a numeric vector of length equal to the number of columns of x}
}
\value{
The value of center determines how column centering is performed. If center is a numeric vector with length equal to the number of columns of x, then each column of x has the corresponding value from center subtracted from it. If center is TRUE then centering is done by subtracting the column means (omitting NAs) of x from their corresponding columns, and if center is FALSE, no centering is done.

The value of scale determines how column scaling is performed (after centering). If scale is a numeric vector with length equal to the number of columns of x, then each column of x is divided by the corresponding value from scale. If scale is TRUE then scaling is done by dividing the (centered) columns of x by their standard deviations if center is TRUE, and the root mean square otherwise. If scale is FALSE, no scaling is done.
}
\description{
scaledf
}
