% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewn_skewt.R
\name{Skew-Normal Distribution}
\alias{Skew-Normal Distribution}
\alias{dsn}
\alias{psn}
\alias{qsn}
\alias{rsn}
\title{Density  function,  distribution  function,  quantiles  and  random  number  generation  for  the  skew normal (SN) and the extended skew-normal (ESN) distribution.}
\usage{
dsn(x, xi = 0, omega = 1, alpha = 0, tau = 0, dp = NULL,
  log = FALSE)

psn(x, xi = 0, omega = 1, alpha = 0, tau = 0, dp = NULL, engine, ...)

qsn(p, xi = 0, omega = 1, alpha = 0, tau = 0, dp = NULL,
  tol = 1e-08, solver = "NR", ...)

rsn(n = 1, xi = 0, omega = 1, alpha = 0, tau = 0, dp = NULL)
}
\arguments{
\item{x}{vector of quantiles. Missing values (NA’s) andInf’s are allowed.}

\item{xi}{vector of location parameters.}

\item{omega}{vector of scale parameters; must be positive}

\item{alpha}{vector of slant parameter(s);+/- Infis allowed. Withpsn, it must be of length 1 ifengine="T.Owen". With qsn, it must be of length 1}

\item{tau}{a single value representing the ‘hidden mean’ parameter of theESNdistribution; tau=0 (default) corresponds to a SN distribution.}

\item{dp}{a vector of length 3 (in the SN case) or 4 (in the ESN case), whose components represent the individual parameters described above. If dp is specified, the individual parameters cannot be set.}

\item{log}{logical flag used in dsn (defaul tFALSE). When TRUE, the logarithm of the density values is returned.}

\item{engine}{a character string which selects the computing engine; this is either "T.Owen" or "biv.nt.prob", the latter from package mnormt. If tau != 0 or length(alpha)>1, "biv.nt.prob" must be used.
If this argument is missing, a default selection rule is applied.}

\item{...}{additional parameters passed to T.Owen}

\item{p}{vector of probabilities. Missing values (NAs) are allowed}

\item{tol}{a scalar value which regulates the accuracy of the result of qsn, measured on the probability scale.}

\item{solver}{a character string which selects the numerical method used for solving the quan-tile equation; possible options are "NR" (default) and "RFB", described in the ‘Details’ section}

\item{n}{a positive integer representing the sample size.}
}
\value{
density (dsn), probability (psn), quantile (qsn) or random sample (rsn) from the skew-normal dis-tribution with given xi,omega and alpha parameters or from the extended skew-normal if tau!=0
}
\details{
psn and qsn make use of function T.Owen or biv.nt.prob

In qsn, the choice solver="NR" selects the Newton-Raphson method for solving the quantile equation, while optionsolver="RFB" alternates a step of regula falsi with one of bisection.
The "NR" method is generally more efficient, but "RFB" is occasionally required in some problematic cases.
}
\examples{

pdf <- dsn(seq(-3, 3, by=0.1), alpha=3)
cdf <- psn(seq(-3, 3, by=0.1), alpha=3)
q <- qsn(seq(0.1, 0.9, by=0.1), alpha=-2)
r <- rsn(100, 5, 2, 5)
qsn(1/10^(1:4), 0, 1, 5, 3, solver="RFB")

}
\references{
Azzalini, A. (1985). A class of distributions which includes the normal ones.Scand. J. Statist.12,171-178.

Azzalini, A. with the collaboration of Capitanio, A. (2014).The Skew-Normal and Related Families.Cambridge University Press, IMS Monographs series.
}
