% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_name.R
\name{standardize_name}
\alias{standardize_name}
\title{\code{standardize_name()}}
\usage{
standardize_name(name, case_out = "upper")
}
\arguments{
\item{name}{A character vector containing one or more strings to standardize.}

\item{case_out}{The desired case of the output vector ("upper" or "lower"; "upper" is default).}
}
\value{
A character vector.
}
\description{
Standardizes names by removing or transforming various special characters and by eliminating excess white space.
}
\details{
\code{standardize_name()} performs several regular-expression standardization procedures to reduce variability in the formatting of names.
The function:

\itemize{
  \item Removes leading and trailing white space
  \item Converts dashes and forward slashes to spaces
  \item Removes the following characters: . ; ( ) [ ] { } # : _
  \item Removes errant commas at the ends of strings
  \item Converts at signs and ampersands to "at" and "and" (and forces surrounding spaces)
  \item Eliminates excess white spaces (e.g., errant double-spaces)
}
}
\examples{
standardize_name(c(' _MAGNOLIA-&-FIR_ #COMPANY#     L.L.C.,',
                   '{MAGNOLIA} /AND/ (F)(I)(R) [COMPANY]... LLC, ',
                   'SMITH, LILY-MAY A.',
                   'SMITH, LILY MAY A,'),
                 case_out = 'upper')

}
