% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tran}
\alias{tran}
\title{Standard and Poors transition matrix 2017}
\format{Trasnsition Matrix
\describe{
  \item{A}{rating A}
  \item{B}{rating B}
  ...
}}
\usage{
tran(object, horizon = horizon)
}
\description{
Credit migration, or a transition matrix, indicates changes in the quality of settled credit at a
particular company. Transition matrices are the main input in various applications of risk
management.
}
