% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Transition_matrix.R
\docType{class}
\name{Cretit migration}
\alias{Cretit migration}
\alias{read_mat}
\alias{tran}
\alias{abs_pd}
\alias{con_pd}
\alias{cum_pd}
\title{Functions to apply to the credit migration or a transition matrix}
\usage{
mat = read_mat(file = "transition_matrix.xlsx")

read_mat(file = file)

tran(object, horizon = horizon)

abs_pd(object, horizon = horizon)

con_pd(object, horizon = horizon)

cum_pd(object, horizon = horizon)
}
\arguments{
\item{object}{TranMat object}

\item{horizon}{Number of year for the computation}
}
\value{
readmat reads the matrix from a excel file and puts it in the right format

the function 'tran' compute the transtitio matrix to horizon years

the function 'abs_pd' compute the absolute probability matrix of default to horizon years

the function 'con_pd' compute the conditional probability matrix of default to horizon years

the function 'cum_pd' compute the cumulative probability matrix of default to horizon years
}
\details{
Credit migration, or a transition matrix, indicates changes in the quality of settled credit at a
particular company. Transition matrices are the main input in various applications of risk
management.

Credit rating is a process where any credit rating observation can form one of several state
ratings. In this research, it is assumed that the credit rating process follows the Markov chain
process. This means that the probability placed on one state can only be determined by
knowing the state from its previous observation. The assumption of Markov chain in the
credit rating process implies that the credit transition is more time invariant or time
homogenous, where the transition probability remains the same towards time and constant
during the predetermined horizon.
}
\examples{
mat = read_mat(file = file)
tran(mat, horizon =  10)
abs_pd(mat, horizon =  10)
con_pd(mat, horizon =  10)
cum_pd(mat, horizon =  10)

}
\references{
http://www.bis.org/ifc/publ/ifcb31u.pdf
}
