% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Regress.R
\name{weopt}
\alias{weopt}
\title{Optimization weights which maximize the correlation between two matrix}
\usage{
weopt(x = x, y = y)
}
\arguments{
\item{x}{Vector of factors}

\item{y}{Vector of local factors}
}
\value{
weights which maximize the correlation between two matrix
}
\description{
Optimization weights which maximize the correlation between two matrix
}
\examples{
library(ECTools)
X = data.frame(
 x1 = rnorm(100),
 x2 = rnorm(100),
 x3 = rnorm(100)
 )

Y = data.frame(
 y1 = rnorm(100),
 y2 = rnorm(100)
 )

}
