% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_zip.R
\name{extract_zip}
\alias{extract_zip}
\title{\code{extract_zip()}}
\usage{
extract_zip(address, if_multiple = "last")
}
\arguments{
\item{address}{A vector containing one or more address strings from which to extract ZIP Codes. Character or character-coercible.}

\item{if_multiple}{If multiple ZIP Code candidates are detected in a string, which should be returned, the first or the last?
\code{'first'} or \code{'last'}, case-insensitive; \code{'last'} is default.}
}
\value{
A vector.
}
\description{
Extracts five-digit ZIP Codes from addresses.
}
\details{
\code{extract_zip()} extracts five-digit ZIP Codes from addresses stored in a vector.
ZIP Codes in the address vector can either be five-digit or nine-digit; only five-digit ZIP Codes will be returned.
If multiple candidate ZIP Codes are identified for a given address string, the \code{if_multiple} argument governs which is returned,
the first or the last (the last is the default).
}
\examples{
addresses <- c('NORFOLK, VA 23501',
               '100 MAIN ST, ALEXANDRIA, VA 22301',
               '99999 FIRST ST, ROANOKE, VA 24001-9999')
extract_zip(addresses, if_multiple = 'last')

}
