% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxcorlag.R
\name{maxccfdf}
\alias{maxccfdf}
\title{maxccfdf}
\usage{
maxccfdf(x = x, lag.max = 6, allow.negative = T, abs = T)
}
\arguments{
\item{x}{data.frame with data}

\item{lag.max}{maximum lag}

\item{allow.negative}{(Optional) logical, if negative lags are allowed, by default TRUE}
}
\value{
matrix with maxccf
}
\description{
maxccfdf
}
